clinic_ai/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
cryptography==43.0.1
psycopg2-binary==2.9.9
python-jose==3.3.0
jinja2==3.1.4
itsdangerous==2.2.0
python-multipart==0.0.9
authlib==1.3.2
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
dateparser==1.2.0
pydantic==2.9.2
apscheduler==3.10.4
pytest==8.3.3
